Create Mines Road Network question
Mines Road Network

Assume you have a mining company that manages several mines located in different locations. The mines are located around a highway; some are on the north side and some are on the south side. The company wants to connect these different mines with a minimum road network to facilitate its work and minimize costs. Ever mine i has GPS location (LATi,LONGi), the company can build a road between mines i and mine j, the cost of the road is the distance between them if they are on the same side, and the cost is 4*distance if they are on two sides, one in the north and one in the south. Write the implementation of the GetMinRoadsNetworkCost function such that it takes as input:

an integer number N that represents the number of mines, 

2 double arrays that represent the latitude and longitude addresses of each mine (in decimal degrees), and 

the latitude of the line separating the mines north and south (for the sake of simplicity, we are assuming the line is parallel to the earth's equator).

Assume that mines do not have obstacles between them, and you can build straight roads. Also, assume that you can measure the distance in kilometers between two GPS points using ComputeDistance(lat1,long1, lat2, long2). The function returns the distance in km of a straight line between the two GPS points. The function is ready for you and can be used as is.


Output

Your function should calculate and print the minimum possible cost to build a road network that allows trucks to travel between any two mines in the network.